moved general check funcs into own file
